refactor: 구매 금액 입력받는 메소드와 유효성 검사하는 메소드 Util 클래스 메소드로 이동
