+++ src/Lotto.js
@@ -24,23 +24,6 @@ class Lotto {
     }
   }
 
-  static async userMoney() {
-    const userMoney = await Console.readLineAsync("구입 금액을 입력해주세요.");
-    return userMoney;
-  }
-
-  static async validateUserMoney() {
-    while (true) {
-      try {
-        const money = await Lotto.userMoney();
-        Validation.userMoney(money);
-        return money;
-      } catch (error) {
-        Console.print(error.message);
-      }
-    }
-  }
-
   static createNumber() {
     const numbers = Random.pickUniqueNumbersInRange(1, 45, 6);
     return numbers;
